workshop/agent-ai/demo-autogen/requirements.txt
autogen-agentchat~=0.2
matplotlib 
yfinance
flaml